Add FstImpl and use it to implement Replace
