+++ rustfst/src/algorithms/cache/mod.rs
@@ -5,3 +5,5 @@ mod vector_cache_state;
 pub use self::cache_impl::CacheImpl;
 pub use self::cache_state::CacheState;
 pub use self::vector_cache_state::VectorCacheState;
+
+pub mod fst_impl;
